change to CMake
