+++ doc/note
