+++ gui/emulate_device.py
