+++ gui/notes.md
