+++ gui/qtthreadutils.py
