--- swig/msvc/DAQModule.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26430.6
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "DAQModule", "DAQModule\DAQModule.vcxproj", "{45F6B675-90AB-4F2E-AD9A-4C7A0C8B1F22}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{45F6B675-90AB-4F2E-AD9A-4C7A0C8B1F22}.Debug|x64.ActiveCfg = Debug|x64
-		{45F6B675-90AB-4F2E-AD9A-4C7A0C8B1F22}.Debug|x64.Build.0 = Debug|x64
-		{45F6B675-90AB-4F2E-AD9A-4C7A0C8B1F22}.Debug|x86.ActiveCfg = Debug|Win32
-		{45F6B675-90AB-4F2E-AD9A-4C7A0C8B1F22}.Debug|x86.Build.0 = Debug|Win32
-		{45F6B675-90AB-4F2E-AD9A-4C7A0C8B1F22}.Release|x64.ActiveCfg = Release|x64
-		{45F6B675-90AB-4F2E-AD9A-4C7A0C8B1F22}.Release|x64.Build.0 = Release|x64
-		{45F6B675-90AB-4F2E-AD9A-4C7A0C8B1F22}.Release|x86.ActiveCfg = Release|Win32
-		{45F6B675-90AB-4F2E-AD9A-4C7A0C8B1F22}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
