+++ swig/python/test.py
@@ -1,0 +1,1 @@
+import daq
