+++ test/dat/v1.dat
@@ -1,0 +1,2 @@
+0000000 95c3 4db3 95c4 4db3 95c5 4db3 95c6 95c3 4db3 95c4 4db3 95c5 4db3 95c6 95c3 4db3 95c4 4db3 95c5 4db3 95c6 95c3 4db3 95c4 4db3 95c5 4db3 95c6 95c3 4db3 95c4 4db3 95c5 4db3 95c6 95c3 4db3 95c4 4db3 95c5 4db3 95c6 95c3 4db3 95c4 4db3 95c5 4dba
+0000010 95c3 4db3 95c4 4db3 95c5 4db3 95c6 4db3
